Added user Auth and logout to NavBar
