+++ src/components/auth/Login.jsx
@@ -2,7 +2,7 @@ import React, { useState } from "react"
 import { Link } from "react-router-dom"
 import { useNavigate } from "react-router-dom"
 import "./Login.css"
-import { getUserByEmail } from "../../services/userService"
+import { getUserByEmail } from "../../services/userServices.jsx"
 
 export const Login = () => {
   const [email, set] = useState("hpassfield7@netvibes.com")
