+++ src/components/auth/Register.jsx
@@ -1,7 +1,7 @@
 import { useState } from "react"
 import { useNavigate } from "react-router-dom"
 import "./Login.css"
-import { createUser, getUserByEmail } from "../../services/userService"
+import { createUser, getUserByEmail } from "../../services/userServices.jsx"
 
 export const Register = (props) => {
   const [customer, setCustomer] = useState({
