+++ src/components/navBar/navBar.jsx
@@ -1,23 +1,41 @@
 // ssrc/components/navBar/navBar.jsx
 
-import { Link } from "react-router-dom"
-import './navBar.css';
+import { Link, useNavigate } from "react-router-dom";
+import "./navBar.css";
 
 export const NavBar = () => {
-    return (
-        <ul className="navbar">
-            <li className="navbar-item">
-                <Link to="/">Home</Link>
-            </li>
-            <li className="navbar-item">
-                <Link to="/tickets">Tickets</Link>
-            </li>
-            <li className="navbar-item">
-                <Link to="/customers">Customers</Link>
-            </li>
-            <li className="navbar-item">
-                <Link to="/employees">Employees</Link>
-            </li>
-        </ul>
-    );
+const navigate = useNavigate()
+
+  return (
+    <ul className="navbar">
+      <li className="navbar-item">
+        <Link to="/">Home</Link>
+      </li>
+      <li className="navbar-item">
+        <Link to="/tickets">Tickets</Link>
+      </li>
+      <li className="navbar-item">
+        <Link to="/customers">Customers</Link>
+      </li>
+      <li className="navbar-item">
+        <Link to="/employees">Employees</Link>
+      </li>
+      {localStorage.getItem("honey_user") ? (
+        <li className="navbar-item navbar-logout">
+          <Link
+            className="navbar-link"
+            to=""
+            onClick={() => {
+              localStorage.removeItem("honey_user");
+              navigate("/", { replace: true });
+            }}
+          >
+            Logout
+          </Link>
+        </li>
+      ) : (
+        ""
+      )}
+    </ul>
+  );
 };
