+++ src/services/userService.js
@@ -1,15 +1,1 @@
-export const getUserByEmail = (email) => {
-  return fetch(`http://localhost:8088/users?email=${email}`).then((res) =>
-    res.json()
-  )
-}
-
-export const createUser = (customer) => {
-  return fetch("http://localhost:8088/users", {
-    method: "POST",
-    headers: {
-      "Content-Type": "application/json",
-    },
-    body: JSON.stringify(customer),
-  }).then((res) => res.json())
-}
+// this page was put here by the script, it is blank
